+++ source/js/account/user/clientconfig/getClientConfig.ts
@@ -1,7 +1,8 @@
 import {fetchWithUserInfo} from '../../fetchWithUserInfo';
-import {getClientConfigBlockingUpdateTTL, getClientConfigTTL} from '../../../intervals';
+import {getClientConfigBlockingUpdateTTL, getClientConfigTTL, getIdleThreshold} from '../../../intervals';
 import {getCacheAge} from '../../../tools/getCacheAge';
 import {triggerPromise} from '../../../tools/triggerPromise';
+import {getElapsedMillisecondsSinceLastActivity} from '../../../tools/activity';
 import {clientConfigStore} from './storedClientConfig';
 import type {ChangeServerConfig, ClientConfig, ClientConfigCache, RatingSettings} from './storedClientConfig';
 
@@ -26,9 +27,11 @@ const getClientConfig = async () => {
 	}
 
 	const age = getCacheAge(cache);
+	const idleDuration = await getElapsedMillisecondsSinceLastActivity();
+	const freshnessThreshold = getClientConfigTTL() * (idleDuration > getIdleThreshold() ? 2 : 1);
 
 	// Use cache if fresh
-	if (cache && age < getClientConfigTTL()) {
+	if (cache && age < freshnessThreshold) {
 		return cache.value;
 	}